refactor(executor): move linux_sysctl into executor helpers
The sysctl read-then-write-with-sudo primitive is not walltime-specific;
memory mode needs it too.
