+++ src/executor/wall_time/profiler/perf/mod.rs
@@ -8,13 +8,13 @@ use crate::executor::helpers::detect_executable::command_has_executable;
 use crate::executor::helpers::env::is_codspeed_debug_enabled;
 use crate::executor::helpers::env::suppress_go_perf_unwinding_warning;
 use crate::executor::helpers::harvest_perf_maps_for_pids::harvest_perf_maps_for_pids;
+use crate::executor::helpers::linux_sysctl::ensure_linux_profiling_sysctls;
 use crate::executor::helpers::run_with_sudo::wrap_with_sudo;
 use crate::executor::shared::fifo::FifoBenchmarkData;
 use crate::executor::wall_time::profiler::NO_BENCHMARKS_DETECTED_WARNING;
 use crate::executor::wall_time::profiler::Profiler;
 use crate::executor::wall_time::profiler::SAMPLING_RATE_HZ;
 use crate::executor::wall_time::profiler::WALLTIME_METADATA_CURRENT_VERSION;
-use crate::executor::wall_time::profiler::linux_sysctl::ensure_linux_profiling_sysctls;
 use crate::executor::wall_time::profiler::perf::perf_executable::get_working_perf_executable;
 use crate::prelude::*;
 use crate::system::SystemInfo;
