+++ src/executor/wall_time/profiler/samply/mod.rs
@@ -4,10 +4,10 @@ use crate::cli::InternalCommands;
 use crate::cli::samply::SamplyArgs;
 use crate::executor::ExecutorConfig;
 use crate::executor::helpers::command::CommandBuilder;
+use crate::executor::helpers::linux_sysctl::ensure_linux_profiling_sysctls;
 use crate::executor::helpers::run_with_sudo::wrap_with_sudo;
 use crate::executor::shared::fifo::FifoBenchmarkData;
 use crate::executor::wall_time::profiler::Profiler;
-use crate::executor::wall_time::profiler::linux_sysctl::ensure_linux_profiling_sysctls;
 use crate::prelude::*;
 use crate::system::SystemInfo;
 use async_trait::async_trait;
